Move schema validator to separate subpackage
