+++ tests/unit/schema/test_schemas.py
@@ -621,11 +621,11 @@ class TestSchemaValidate(object):
         u('2018-01-02T23:59:59Z'),
     ])
     @mock.patch(
-        'openapi_core.schema.schemas._format.'
+        'openapi_core.schema_validator._format.'
         'DATETIME_HAS_STRICT_RFC3339', True
     )
     @mock.patch(
-        'openapi_core.schema.schemas._format.'
+        'openapi_core.schema_validator._format.'
         'DATETIME_HAS_ISODATE', False
     )
     def test_string_format_datetime_strict_rfc3339(self, value):
@@ -640,11 +640,11 @@ class TestSchemaValidate(object):
         u('2018-01-02T23:59:59Z'),
     ])
     @mock.patch(
-        'openapi_core.schema.schemas._format.'
+        'openapi_core.schema_validator._format.'
         'DATETIME_HAS_STRICT_RFC3339', False
     )
     @mock.patch(
-        'openapi_core.schema.schemas._format.'
+        'openapi_core.schema_validator._format.'
         'DATETIME_HAS_ISODATE', True
     )
     def test_string_format_datetime_isodate(self, value):
